requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
python-multipart
itsdangerous
jinja2
zipstream-new

